front: rename manageTrainScheduleMap
