+++ front/src/modules/trainschedule/components/ManageTrainSchedule/ManageTimetableItemMap/OperationalPointPopupDetails.tsx
